[PT-D][Tensor Parallelism][2/N] Sync TP API change to PT prod
