+++ test/distributed/_tensor/parallel/test_tp_examples.py
@@ -11,18 +11,14 @@ from torch.testing._internal.distributed._tensor.common_dtensor import (
     skip_unless_torch_gpu,
 )
 from torch.distributed._tensor import (
-    distribute_module,
     DeviceMesh,
     Replicate,
 )
 from torch.distributed._tensor.parallel import (
+    PairwiseParallel,
     TensorParallelMultiheadAttention,
-    tp_shard_self_attn,
-    replicate_input,
-    replicate_output,
+    parallelize_module,
 )
-from torch.distributed._tensor.parallel import PairwiseParallel
-from torch.distributed._tensor.parallel.api import _parallelize_mlp
 
 
 class MLPModule(torch.nn.Module):
@@ -70,7 +66,7 @@ class DistTensorParallelExampleTest(DTensorTestBase):
             self.device_type,
             torch.arange(0, NUM_DEVICES),
         )
-        _parallelize_mlp(model_tp, device_mesh, PairwiseParallel())
+        model_tp = parallelize_module(model_tp, device_mesh, PairwiseParallel())
         optim = torch.optim.SGD(model.parameters(), lr=LR)
         optim_tp = torch.optim.SGD(model_tp.parameters(), lr=LR)
 
@@ -182,13 +178,7 @@ class DistTensorParallelExampleTest(DTensorTestBase):
 
         # Shard module and initialize optimizer.
         device_mesh = DeviceMesh(self.device_type, list(range(NUM_DEVICES)))
-        distribute_module(
-            model_tp,
-            device_mesh,
-            partition_fn=tp_shard_self_attn,
-            input_fn=replicate_input,
-            output_fn=replicate_output,
-        )
+        parallelize_module(model_tp, device_mesh, PairwiseParallel())
 
         device_mesh = model_tp.qkv.weight.device_mesh
         replicate = [Replicate()] * device_mesh.ndim
@@ -339,13 +329,7 @@ class DistTensorParallelExampleTest(DTensorTestBase):
 
         # Shard module and initialize optimizer.
         device_mesh = DeviceMesh(self.device_type, list(range(NUM_DEVICES)))
-        distribute_module(
-            model_tp,
-            device_mesh,
-            partition_fn=tp_shard_self_attn,
-            input_fn=replicate_input,
-            output_fn=replicate_output,
-        )
+        parallelize_module(model_tp, device_mesh, PairwiseParallel())
 
         device_mesh = model_tp.attn.qkv.weight.device_mesh
         replicate = [Replicate()] * device_mesh.ndim
